board/serializers.py
from rest_framework import serializers # serializer import
from .models import * # 선언한 모델 import

class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = '__all__'


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'